Changes: - Added Token splitting protocol to missing services. - Bugfix: the remaining tokens are persisted in a Singleton structure, instead of the scoped metadata structure. This is to fix an issue where changes made to the scoped data in the HTTP Handler (TCCHttpInjector) were not reflected when the context was back in the TCC Middleware layer. - Added tokens to read-based services. All executed services send their tokens to the coordinator. This is to remove the initial assumption that we knew if a service executed read-only transactions.

+++ src/Services/Catalog/Catalog.API/DependencyServices/IScopedMetadata.cs
@@ -1,6 +1,6 @@
 ﻿namespace Microsoft.eShopOnContainers.Services.Catalog.API.DependencyServices;
 public interface IScopedMetadata {
-    double Tokens { get; set; }
+    int Tokens { get; set; }
     DateTime Timestamp { get; set; }
     string ClientID { get; set; }
     bool ReadOnly { get; set; }
