+++ src/Services/Discount/Discount.API/DependencyServices/ScopedMetadata.cs
@@ -1,7 +1,7 @@
 ﻿namespace Microsoft.eShopOnContainers.Services.Discount.API.DependencyServices;
 
 public class ScopedMetadata : IScopedMetadata {
-    double tokens;
+    int tokens;
     // The default timestamp should not be used by any functionality. This is used mainly for ContextSeed population.
     DateTime timestamp = DateTime.UtcNow;
     string clientID;
@@ -9,7 +9,7 @@ public class ScopedMetadata : IScopedMetadata {
 
     public ScopedMetadata() { }
 
-    public double Tokens {
+    public int Tokens {
         get { return tokens; }
         set { tokens = value; }
     }
